// src/app/modules/shares/models/request/import-evoluation-req.model.ts
export interface ImportEvoluationReqModel {
  org_unit_code: string;
  service_work_score: number;
  investigative_work_score: number;
  crime_prevention_work_score: number;
  traffic_work_score: number;
  satisfaction_score: number;
  service_work_count: number;
  investigative_work_count: number;
  crime_prevention_work_count: number;
  traffic_work_count: number;
  satisfaction_count: number;
  evaluators_amount: number;
  evaluation_year: string;
}